9pfs: disallow the NUL character in all strings
According to the 9P spec at http://man.cat-v.org/plan_9/5/intro :

Data items of larger or variable lengths are represented by a
two-byte field specifying a count, n, followed by n bytes of
data.  Text strings are represented this way, with the text
itself stored as a UTF-8 encoded sequence of Unicode charac-
ters (see utf(6)). Text strings in 9P messages are not NUL-
terminated: n counts the bytes of UTF-8 data, which include
no final zero byte.  The NUL character is illegal in all
text strings in 9P, and is therefore excluded from file
names, user names, and so on.

With this patch, if a 9P client sends a text string containing a NUL
character, the request will fail and the client is returned EINVAL.
If QEMU does the same, then it is a bug that needs to be fixed and QEMU
will abort.

The checking is done in the marshalling code because it is a convenient
place to check all strings.

The proxy backend uses V9fsString to exchange path names and extended
attribute values with the proxy helper. In both cases, we can have
zeroes in the buffer: the proxy marshalling code is then updated to
have its own implementation of the 's' format, that does not care
for NUL characters.

Suggested-by: Peter Maydell <peter.maydell@linaro.org>
Signed-off-by: Greg Kurz <groug@kaod.org>