9pfs: delay PDUs that could fail during migration
Some 9P operations may cause a fid to become unreclaimable, in which
case they try to add a migration blocker. If a migration is already
in progress, the blocker cannot be added, the 9P operation fails and
the EBUSY error is returned to the guest.

This patch introduces a list to park PDUs that could fail, instead
of actually submitting them. Such PDUs are still considered active
but don't cause any I/O, causing the client to simply wait for the
answer. When it is time to migrate the device state, we put all
these not-yet-started requests back to the transport vring, so that
they'll be migrated with the guest memory and be submitted again
when migration is over.

Signed-off-by: Greg Kurz <groug@kaod.org>