9pfs: add migration support
This adds VMState based migration support for 9pfs devices. It relies
on the reclaim fd mechanism, to close open fds before migration on the
source and reopen it after migration (on the destination in case of
sucess, or on the source otherwise). It is hence assumed that we don't
have any unreclaimable fid at migration time.

Signed-off-by: Greg Kurz <groug@kaod.org>