migration: introduce VMS_VBUFFER_UINT32
The VMS_VBUFFER flag assumes that entry size is stored in an int32_t.
Similarly to what we already do for VMS_VARRAY, this patch introduces
an UINT32 variant, which assumes the size is stored in an uint32_t.

The VMSTATE_VBUFFER_UINT32() and VMSTATE_VBUFFER_ALLOC_UINT32() macros
are converted to use this flag, since they're already uint32_t based.

Signed-off-by: Greg Kurz <groug@kaod.org>