migration: convert vmstate_size() to size_t
This function is expected to return a number of bytes, which is either
directly passed to g_malloc() or to VMStateInfo::get() and put() ops.
It doesn't return any error that would justify to use a signed type.

To make this clear, this patch converts vmstate_size() to size_t.

Signed-off-by: Greg Kurz <groug@kaod.org>