9pfs: add migration blocker to unreclaimable fids
A fid becomes unreclaimable if we know we cannot re-open its associated
file. This can happen either if the file gets unlinked or if the O_EXCL
flag was used to open the file.

Future migration support will need to close/re-open open fids. This
means that unreclaimable fids cannot be migrated.

This patch ensures the migration is blocked as long as we have an
unreclaimable fid around. This is done by reusing the current
blocker from the session with a reference counter.

If the global migration.only-migratable property is true, then the
affected client operations will fail if we cannot set the blocker
(ie, a migration is in progress).

Signed-off-by: Greg Kurz <groug@kaod.org>