9pfs: drain requests before migration
We cannot capture the state of an active PDU. To allow migration, we
have to wait for active PDUs to complete and push their result to the
client. Since completion happens in the global aio context, we can
simply loop around aio_poll() until the active list is empty (as we
already do it at reset time).

Signed-off-by: Greg Kurz <groug@kaod.org>