spapr/nvlink2: Rework getting of "nvlink2-*" properties
VFIO configures some quirks for NVLINK2. This means that some
properties are optional but if they do exist they should be
gettable with the appropriate type.

Use object_property_find() to know if a property exists and
if we're sure that the property exists, pass &error_abort when
getting it.

Note that the "size" property of the NVLINK2 memory regions is
always gettable in spapr_phb_nvgpu_ram_populate_dt() since
this function is called at reset time.

Signed-off-by: Greg Kurz <groug@kaod.org>