virtio-mem-pci: Pass &error_abort when getting VIRTIO_MEM_SIZE_PROP
The VIRTIO_MEM_ADDR_PROP property is added to every instance of the
virtio-mem device class in virtio_mem_instance_init():

    object_property_add(obj, VIRTIO_MEM_SIZE_PROP, "size", virtio_mem_get_size,
                        NULL, NULL, NULL);

It is then added to every instance of the virtio-mem-pci device class as
an alias in virtio_mem_pci_instance_init():

    object_property_add_alias(obj, VIRTIO_MEM_SIZE_PROP, OBJECT(&dev->vdev),
                              VIRTIO_MEM_SIZE_PROP);

virtio_mem_get_size() relies on visit_type_size() which never fails. It
is thus expected that the property is always gettable.

Signed-off-by: Greg Kurz <groug@kaod.org>