hw/arm/bcm2835: Pass &error_abort when getting the "vcram-size" property
bcm2835_peripherals_realize() retrieves the "vcram-size" property of
the BCM2835 peripherals device by calling object_property_get_uint()
with an &err argument. It simply propagates the error to the caller
in case of failure.

This property comes from the default property list of the BCM2835
frame buffer device:

    DEFINE_PROP_UINT32("vcram-size", BCM2835FBState, vcram_size,
                       DEFAULT_VCRAM_SIZE),

It is added to every instance of the BCM2835 peripherals device as
an alias in bcm2835_peripherals_init():

    /* Framebuffer */
    object_initialize_child(obj, "fb", &s->fb, TYPE_BCM2835_FB);
    object_property_add_alias(obj, "vcram-size", OBJECT(&s->fb), "vcram-size");

It is thus safe to assume that it is always gettable.

Signed-off-by: Greg Kurz <groug@kaod.org>