hostmem: Use object_property_find() for "mem-path"
The goal of host_memory_backend_pagesize() is to return the underlying
host page size of a memory device. This is usually the default host page
size, unless the memory device is backed by huge pages, in which case the
size can be retrieved with statfs().

Since only file backed memory devices can bring huge pages, we get their
"mem-path" property and pass it to qemu_mempath_getpagesize() which calls
statfs() when passed a non-null argument or returns the default host page
size otherwise.

The "mem-path" property is retrieved by calling object_property_get_str()
with a NULL errp. This works well but ignoring errors isn't really a
recommanded practice.

Instead of that, first check the existence of "mem-path" with
object_property_find() and pass &error_abort to object_property_get_str()
if it's there.

Signed-off-by: Greg Kurz <groug@kaod.org>