hw/mem/nvdimm: Use object_property_find() when getting "pmem"
An NVDIMM device might be backed by persistent memory, ie. a
file, or not, ie. RAM. This can be checked through the "pmem"
property, which is only created by the hostmem file backend
class in its class init function:

    object_class_property_add_bool(oc, "pmem",
        file_memory_backend_get_pmem, file_memory_backend_set_pmem);

file_memory_backend_get_pmem() cannot fail. It is thus safe to assume
that "pmem" is gettable when present.

Signed-off-by: Greg Kurz <groug@kaod.org>