hw/acpi/nvdimm: Pass &error_abort when getting PC_DIMM_SIZE_PROP
Both nvdimm_build_structure_spa() and nvdimm_build_structure_memdev()
retrieve the size of the DIMM by calling object_property_get_uint()
on PC_DIMM_SIZE_PROP with a NULL errp argument. This means that a
programming error (eg. the property can't be retrieved) will be
treated like the size of the DIMM is zero, which doesn't make sense.

Since both functions are called on the memory plug path only if
memory_device_pre_plug() previously succeeded. It is thus safe to
assume that the memory region is present and that we can get its
size through PC_DIMM_SIZE_PROP.

Signed-off-by: Greg Kurz <groug@kaod.org>