hw/acpi/memory_hotplug: Pass &error_abort when getting PC_DIMM_SIZE_PROP
acpi_memory_hotplug_read() retrieves the size of the DIMM device by
calling object_property_get_uint() on PC_DIMM_SIZE_PROP with a NULL
errp argument. This means that a programming error (eg. the property
can't be retrieved) will be treated like the size of the DIMM is zero,
which doesn't make sense.

Since acpi_memory_plug_cb() sets mdev->dimm to point to a realized
DIMM device, it is safe to assume that the associated memory region
exists and that we can get its size through PC_DIMM_SIZE_PROP when
mdev->dimm is set.

Signed-off-by: Greg Kurz <groug@kaod.org>