qdev-monitor: Simplify handling of link properties in 'info qtree'
We don't show link properties in 'info qtree' but this happens
in a rather convoluted manner. We first try to check if an
associated legacy property exists, which is the case for link
properties since qdev_prop_link doesn't implement .print nor
.get. Then we try to the retrieve the stringified value of the
property with object_property_get_str(), which fails and returns
NULL since the object property doesn't have a get method. Finally
we skip the printing since value is NULL. Icing on the cake, all
of this relies on ignoring errors of object_property_get_str()
and object_property_print() which isn't a recommanded practice.

Instead of that, register link properties as legacy properties like
we already do for pointer properties, and pass &error_abort to
both object_property_get_str() and object_property_print() because
a failure there is necessarily a programming error.

Signed-off-by: Greg Kurz <groug@kaod.org>